[Close #214] Move User::New action.
`Newpoc::Action::User::New` replaced by `UsersController::Action::New`.

RSpec: 725 examples, 0 failures
RuboCop: 246 files inspected, no offenses detected
