--- lib/newpoc-action-user-new/.rspec
@@ -1,2 +1,0 @@
---format documentation
---color
