--- lib/newpoc-action-user-new/Gemfile
@@ -1,4 +1,0 @@
-source 'https://rubygems.org'
-
-# Specify your gem's dependencies in newpoc-action-user-new.gemspec
-gemspec
