--- lib/newpoc-action-user-new/lib/newpoc/action/user/new/version.rb
@@ -1,11 +1,0 @@
-
-module Newpoc
-  module Action
-    module User
-      # New-user controller-action encapsulation for `new_poc`.
-      class New
-        VERSION = '0.0.1'
-      end
-    end
-  end
-end
