--- lib/newpoc-action-user-new/spec/spec_helper.rb
@@ -1,2 +1,0 @@
-$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'newpoc/action/user/new'
